Create hashing algorithms
This video speaks about how the hashing algorithms play an essential role in our society. Hashing is very popular in the cryptocurrency craze we live through today because it allows the transfer of information through a secure and fast means. Hashing allows our systems to map data and access it at an incredible pace.